+++ src/ParallaxGenDirectory.cpp
@@ -33,7 +33,7 @@ void ParallaxGenDirectory::findHeightMaps(const vector<wstring> &Allowlist, cons
   // find height maps
   spdlog::info("Finding parallax height maps");
   // Find heightmaps
-  HeightMaps = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true);
+  HeightMaps = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true, false);
   sort(HeightMaps.begin(), HeightMaps.end());
   spdlog::info("Found {} height maps", HeightMaps.size());
 }
@@ -43,7 +43,7 @@ void ParallaxGenDirectory::findComplexMaterialMaps(const vector<wstring> &Allowl
   spdlog::info("Finding complex material maps");
   // find complex material maps
   // TODO find a way to have downstream stuff edit this directly instead of replacement
-  ComplexMaterialMaps = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, false);
+  ComplexMaterialMaps = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, false, false);
   sort(ComplexMaterialMaps.begin(), ComplexMaterialMaps.end());
   // No conclusion because this is affected by D3D later
 }
@@ -52,7 +52,7 @@ void ParallaxGenDirectory::findMeshes(const vector<wstring> &Allowlist, const ve
                                       const vector<wstring> &ArchiveBlocklist) {
   // find Meshes
   spdlog::info("Finding Meshes");
-  Meshes = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true);
+  Meshes = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true, true);
   spdlog::info("Found {} Meshes", Meshes.size());
 }
 
@@ -61,7 +61,7 @@ void ParallaxGenDirectory::findTruePBRConfigs(const vector<wstring> &Allowlist, 
   // TODO more logging here
   // Find True PBR Configs
   spdlog::info("Finding TruePBR Configs");
-  auto ConfigFiles = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true);
+  auto ConfigFiles = findFiles(true, Allowlist, Blocklist, ArchiveBlocklist, true, false);
 
   // loop through and parse Configs
   size_t ConfigOrder = 0;
@@ -78,6 +78,8 @@ void ParallaxGenDirectory::findTruePBRConfigs(const vector<wstring> &Allowlist, 
         if (Element.contains("texture")) {
           Element["match_diffuse"] = Element["texture"];
         }
+
+        Element["json"] = Config.string();
 
         // loop through filename Fields
         for (const auto &Field : getTruePBRConfigFilenameFields()) {